refactor: split TaskSidebarRows.tsx into focused row components
- Extract TaskGroupRow.tsx (includes shared TreeTitleCell + ContextMenuTarget type)
- Extract SortableTaskRow.tsx
- Update imports in TaskSidebar.tsx and TaskSidebarContextMenu.tsx
- Delete monolithic TaskSidebarRows.tsx (617 -> ~320 + ~312 LOC)
- Add test plan doc per project guidelines

This improves separation of concerns and file focus (RULES.md).
No behavior changes; all tests, type-check, and build pass.

Pre/post verification completed per TASKSIDEBAR_ROWS_REFACTOR_PLAN.md
